+++ src/stores/signUp.ts
@@ -1,0 +1,18 @@
+import { atom } from 'recoil';
+
+export interface SignUpFormTypes {
+  id: string;
+  password: string;
+  email: string;
+  nickname: string;
+}
+
+export const SignUpFormErrorState = atom({
+  key: 'SignUpFormErrorState',
+  default: true,
+});
+
+export const SignUpFormState = atom<SignUpFormTypes>({
+  key: 'SignUpFormState',
+  default: { id: '', password: '', email: '', nickname: '' },
+});
